backend/requirements.txt
# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.0.0

# Data & ML
numpy>=1.26.0
pandas>=2.0.0
torch>=2.0.0
sentence-transformers>=3.0.0
scipy>=1.10.0

# Polymarket / baskets
requests>=2.28.0

# LLM (basket-llm-no-target, create_basket_llm)
openai>=1.0.0
python-dotenv>=1.0.0

# Visualization (create_basket)
matplotlib>=3.7.0
seaborn>=0.12.0

# Graph (optional)
neo4j>=5.0.0
spacy>=3.7.0
